requirements.txt
plaid-python>=20.0.0
pyyaml>=6.0
apscheduler>=3.10,<4
cryptography>=42.0
pytest>=7.0
